Move yfinance batch download function to scraping utils
